refactor timer into /game/handler
